backend/requirements.txt
flask
librosa
numpy
scipy
torch
torchvggish
faiss-cpu 